refactor: integration router refactored
